Move types from SquatchPortal.ts to environment.ts
