+++ packages/component-boilerplate/src/environment/environment.ts
@@ -1,4 +1,20 @@
-import { PortalEnv } from "./SquatchPortal";
+/**
+ * Environment provided by components hosted in a web component (`sqh-widget`)
+ */
+
+declare global {
+  interface Window {
+    SquatchPortal?: PortalEnv;
+  }
+}
+
+/**
+ * Portal env doesn't include User Id
+ */
+export type PortalEnv = Pick<
+  WidgetIdent,
+  "tenantAlias" | "appDomain" | "programId"
+> & { portalAuthUrl: string };
 
 /**
  * Program ID context helpers
